requirements.txt
python-dotenv
asyncio_redis
aiofiles==0.4.0
aiomysql==0.0.20
asn1crypto==0.24.0
cffi==1.12.2
cryptography==2.6.1
httptools==0.0.13
Jinja2==2.10
MarkupSafe==1.1.1
multidict==4.5.2
pycparser==2.19
PyMySQL==0.9.2
sanic==18.12.0
sanic-jinja2==0.7.4
sanic-mysql==0.0.5
six==1.12.0
websockets==6.0